CLUTf32 value used a wrong data type enum

+++ ColorManagerTests/ICC/Data/LUTData.cs
@@ -163,7 +163,7 @@ namespace ColorManagerTests.ICC.Data
                 new double[] { 4f, 5f, 6f },
                 new double[] { 7f, 8f, 9f },
             },
-            2, 3, new byte[] { 3, 3 }, CLUTDataType.UInt16
+            2, 3, new byte[] { 3, 3 }, CLUTDataType.Float
         );
 
         /// <summary>
